Moved migration and implemented on import
This is a dirty fix that enables compatibility with our backup/restore system in it's current state.

Every migration 9+ will need to also be implemented in the Import container like this commit shows. Each migration will also have to be intelligent enough to not reapply itself since it can be run against a backup from any version 9+.